Fix IB in pointer cast to larger type
https://pvs-studio.com/en/docs/warnings/v1091/print/

To convert a 32-bit pointer to a 64-bit integer and avoid
the implementation-defined behavior, do the following:
* Convert a 32-bit pointer to a 32-bit integer.
* Convert the resulting 32-bit integer to a 64-bit integer.

+++ vphysics_jolt/vjolt_object.cpp
@@ -40,7 +40,7 @@ JoltPhysicsObject::JoltPhysicsObject( JPH::Body *pBody, JoltPhysicsEnvironment *
 	static_assert( offsetof( JoltPhysicsObject, m_pGameData ) == sizeof( void * ) );
 
 	// Set the body's userdata as ourselves
-	pBody->SetUserData( reinterpret_cast<uint64>( this ) );
+	pBody->SetUserData( static_cast<uint64>( reinterpret_cast<uintptr_t>(this) ) );
 	if ( !m_pBody->IsStatic() )
 	{
 		JPH::MotionProperties* pMotionProperties = m_pBody->GetMotionProperties();
@@ -1178,7 +1178,7 @@ void JoltPhysicsObject::RestoreObjectState( JPH::StateRecorder &recorder )
 	m_pBody->RestoreState( recorder );
 
 	// Set the body's userdata as ourselves
-	m_pBody->SetUserData( reinterpret_cast<uint64>( this ) );
+	m_pBody->SetUserData( static_cast<uint64>( reinterpret_cast<uintptr_t>(this) ) );
 
 	// Josh: Do not read m_pGameData, as this is not serialized.
 	//recorder.Read( m_pGameData );
